refactor: 코틀린 컨벤션에 맞게 수정
- 대문자 패키지명 Information -> info 로 수정
