+++ app/src/main/java/com/project/meongcare/info/model/data/remote/ProfileApi.kt
@@ -1,6 +1,6 @@
-package com.project.meongcare.Information.model.data.remote
+package com.project.meongcare.info.model.data.remote
 
-import com.project.meongcare.Information.model.entities.GetDogInfoResponse
+import com.project.meongcare.info.model.entities.GetDogInfoResponse
 import com.project.meongcare.home.model.entities.GetDogListResponse
 import com.project.meongcare.home.model.entities.GetUserProfileResponse
 import okhttp3.MultipartBody
