+++ app/src/main/java/com/project/meongcare/info/view/ProfileDogAdapter.kt
@@ -1,9 +1,8 @@
-package com.project.meongcare.Information.view
+package com.project.meongcare.info.view
 
 import android.content.Context
 import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.cardview.widget.CardView
